Task 11 Retry

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -40,7 +40,7 @@ void largest_prime_factor(long int n)
  */
 int is_prime(long int c)
 {
-	int range;
+	long int range;
 
 	for (range = 2 ; range < c ; range++)
 	{
